backend/requirements.txt
# Web Framework and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# Caching and Rate Limiting
slowapi==0.1.9
cachetools==5.3.2

# Configuration
python-dotenv==1.0.0

# Additional utilities
python-dateutil==2.8.2
requests==2.31.0
